requirements.txt
absl-py==1.0.0
APScheduler==3.6.3
arch==5.3.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.1
cycler==0.11.0
fonttools==4.29.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
patsy==0.5.3
Pillow==9.0.1
property-cached==1.6.4
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
statsmodels==0.13.5
threadpoolctl==3.1.0
torch==1.10.2
tornado==6.1
tqdm==4.63.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
